pyproject.toml
[tool.poetry]
name = "cosmosdb"
version = "0.1.0"
description = ""
authors = ["Laurie Voss <github@seldo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.9.25.post1"
python-dotenv = "^1.0.0"
pymongo = "^4.6.1"
flask = "^3.0.0"
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
